backend/pyproject.toml
[project]
name = "wm-system-backend"
version = "0.1.0"
description = "Inventory management app"
requires-python = "==3.12.*"
dependencies = [
    "asgiref==3.8.1",
    "attrs==25.3.0",
    "autopep8==2.3.2",
    "certifi==2025.4.26",
    "chardet==5.2.0",
    "charset-normalizer==3.4.2",
    "colorama==0.4.6",
    "diff-match-patch==20241021",
    "django==5.2",
    "django-admin-rangefilter==0.13.2",
    "django-cors-headers==4.7.0",
    "django-import-export==4.3.7",
    "django-silk==5.3.2",
    "djangorestframework==3.16.0",
    "drf-spectacular==0.28.0",
    "drf-spectacular-sidecar==2025.5.1",
    "et-xmlfile==2.0.0",
    "google-cloud-storage==3.1.0",
    "gprof2dot==2025.4.14",
    "gunicorn==23.0.0",
    "idna==3.10",
    "inflection==0.5.1",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2025.4.1",
    "numpy==2.2.5",
    "openpyxl==3.1.5",
    "packaging==25.0",
    "pandas==2.2.3",
    "pillow==11.2.1",
    "pip==25.1.1",
    "psycopg2-binary==2.9.10",
    "pycodestyle==2.13.0",
    "python-dateutil==2.9.0.post0",
    "python-decouple==3.8",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "qrcode==8.2",
    "referencing==0.36.2",
    "reportlab==4.4.0",
    "requests==2.32.3",
    "rpds-py==0.24.0",
    "setuptools==80.3.1",
    "six==1.17.0",
    "sqlparse==0.5.3",
    "tablib==3.8.0",
    "typing-extensions==4.13.2",
    "tzdata==2025.2",
    "uritemplate==4.1.1",
    "urllib3==2.4.0",
    "wheel==0.45.1",
]
